Updated testDevideByAnInteger signature (changed denominator type from long to double) Updated hyperbola signature (changed x type from long to double) Updated consoleReader: - changed return value type from long to double - in.nextLong replaced on nextDouble for flexibility - added US locale to be able to identify double in the string

+++ app/src/main/java/application/functions/LMath.java
@@ -4,7 +4,7 @@ import application.exception.ZeroValueException;
 
 public final class LMath {
 
-    public static double hyperbola(long x) throws ZeroValueException {
+    public static double hyperbola(double x) throws ZeroValueException {
         if(x == 0){
             throw new ZeroValueException();
         }
